Added logout endpoint

+++ resources/item.py
@@ -6,7 +6,7 @@ from sqlalchemy.exc import SQLAlchemyError
 from db import db
 from models import ItemModel
 from schemas import ItemSchema, ItemUpdateSchema
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 blp = Blueprint("Items", __name__, description="Operations on items")
 
@@ -42,6 +42,11 @@ class Item(MethodView):
                 dict: Response message/data
                 int: The status code of the response
         """
+        
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+        
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
